+++ backend/src/main/java/com/project/Rentingaccommodation/service/UserService.java
@@ -5,11 +5,11 @@ import java.util.List;
 import com.project.Rentingaccommodation.model.User;
 
 public interface UserService {
-	User findOne(Long id);
+	User findOne(Integer id);
 	List<User> findAll();
 	User save(User user);
 	List<User> save(List<User> users);
-	User delete(Long id);
-	void delete(List<Long> ids);
+	User delete(Integer id);
+	void delete(List<Integer> ids);
 	User findByEmail(String email);
 }
